User posts (front and backend), comments, collections retrieving frontend. User update on login through OAuth

+++ pinned-backend/crates/api_routes/src/comments/dto.rs
@@ -21,5 +21,5 @@ pub struct NewCommentDTO {
 #[derive(Serialize)]
 pub struct NewCommentMessage {
     pub message: String,
-    pub comment_id: usize 
+    pub comment_id: i32 
 }
